+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Markdown editor: Support LaTeX math expressions (http://localhost:9090/reporting/markdown-features/#math-expressions)
 * AI agent: Add Mistral AI provider
 * Increase resolution of charts and mermaid diagrams in PDFs
+* Support images in excalidraw notes
 
 
 ## v2025.110 - 2025-12-22